requirements.txt
flasgger==0.9.7.1
Flask==3.1.2
python-dotenv==1.1.1
